==> new_code/04b_lstm_rnn.ipynb <==
## **Notebook Contents**
- Import Libraries
- Import Dataframes
- Word Cleaning
- Preprocess Data
- Modeling
- Scores
- Citations

<a name="importlibrarieml"></a>
## **Import Libraries**

# Standard Imports
import numpy as np 
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import train_test_split, cross_val_score

# NLP Imports
import re
import nltk
nltk.download("stopwords")
from nltk.corpus import stopwords
from nltk import word_tokenize
STOPWORDS = set(stopwords.words('english'))

# Keras Imports
from keras.preprocessing.text import Tokenizer
from keras.preprocessing.sequence import pad_sequences
from keras.models import Sequential
from keras.layers import Dense, Embedding, LSTM, SpatialDropout1D
from sklearn.model_selection import train_test_split
from keras.utils.np_utils import to_categorical
from keras.callbacks import EarlyStopping
from keras.layers import Dropout
import tensorflow as tf
from tensorflow import keras
# Credit to https://towardsdatascience.com/multi-class-text-classification-with-lstm-1590bee1bd17 for main inspiration and code

<a name="importdataframeml"></a>
## **Import Dataframes**


df = pd.read_csv('../data/ready_for_modeling.csv')

zeroes = df[df['label'] == 0]

zeroes['tweet'][1323]

zeroes['tweet'][34]

zeroes["tweet"][32]

df.head()

df.shape

df['tweet'].values

<a name="preprocessml"></a>
### **Preprocessing Data** 

# The maximum number of words to be used. (most frequent)
max_words = 15_000

# Max number of words in each tweet.
# First 280 words in the tweet
max_sequence_length = 280

# This is the second argument in our embedding layer 
embedding_dimensions = 100

# Keras Tokenizer turning each Tweet in the corpus into either a sequence of integers or into a vector
# Instantiate the Tokenizer
tokenizer = Tokenizer(num_words=max_words, filters='!"#$%&()*+,-./:;<=>?@[\]^_`{|}~', lower=True)

# Use the tokenizer on every document in our corpus
tokenizer.fit_on_texts(df['tweet'].values)

# Replaces the word with it's index
word_index = tokenizer.word_index
print('Found %s unique tokens.' % len(word_index))

# Taking the Tweets in df['tweet'] and Tokenizing the list of Tweets
X = tokenizer.texts_to_sequences(df['tweet'].values)

# Keras pad sequence --> Make sequences the same size! Makes the shape the same 
X = pad_sequences(X, maxlen=max_sequence_length) 
print('Shape of data tensor:', X.shape)

# Turn our classes into 0's and 1's
y = pd.get_dummies(df['label']).values
print('Shape of label tensor:', y.shape)

y

# Split our data into training and testing
X_train, X_test, y_train, y_test = train_test_split(X,y, test_size = 0.30, random_state = 42)
print('Train')
print(X_train.shape,y_train.shape) # training data
print('='*40)
print('Test')
print(X_test.shape,y_test.shape) # testing data

len(X)

<a name="modelingml"></a>
## **Modeling**

model = Sequential() #Instantiate the Sequential Model

model.add(Embedding(max_words, embedding_dimensions, input_length=X.shape[1])) # Adding the embedding layer 1st
model.add(SpatialDropout1D(0.10)) # Adding dropout layers to limit overfitting
model.add(LSTM(100, dropout=0.10, recurrent_dropout=0.10)) # Adding dropout layers to limit overfitting
model.add(Dense(2, activation='softmax'))
model.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['accuracy'])

epochs = 5
batch_size = 64

history = model.fit(X_train, y_train, epochs=epochs, batch_size=batch_size,validation_split=0.1,callbacks=[EarlyStopping(monitor='val_loss', patience=3, min_delta=0.001)])

model.summary()

# Evaluating our model on the Testing Data
accr = model.evaluate(X_test,y_test)
print('Test set\n  Loss: {:0.3f}\n  Accuracy: {:0.3f}'.format(accr[0],accr[1]))

# Example of how the model would classify a Tweet
new_tweet = ["I feel like David is going to try to act surprised when he’s saved. Problem is no one will buy it. Everyone suspects he or kevin have a power."]
# The above text shows a misclassification
seq = tokenizer.texts_to_sequences(new_tweet)
padded = pad_sequences(seq, maxlen=max_sequence_length)
pred = model.predict(padded)
labels = [0, 1]
print(pred, labels[np.argmax(pred)])

# Save the entire model as a SavedModel.
model.save('saved_model/lstm_rnn_model_')

print(tf.__version__)
